=== Snapshot Location === Change of goal
Okay so here is where things get odd.

Because data.gov is *so* shit we are just going to import data off http://data.okfn.org/data
and then use that to deal with the data issue, I'm going to keep the history in here but for
now we are mainly interseted in the data play part of the software rather then the importing
process. This means for now I just need to put stuff in SQL tables and then hope for da best