Added owner testing when searching in the dataset index.
However this one is balls slow right now and needs to be sorted.

http://i.imgur.com/UnideQB.png This slow.

Another thing I realised is that if you run this, and want to kill it.
you can't if you stab it with the KILL connection keyword then go will
auto remake the connection and try again. While that may seem like a good
thing that is also a TERRIBLE THING.

I mean holy hell that could really wreck.

The only way you know that it has gone wrong is this:
[MySQL] 2014/02/20 11:46:10 packets.go:30: EOF
[MySQL] 2014/02/20 11:46:10 statement.go:24: Invalid Connection

and then it just tries again. :(